requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==2.2.10
django-crispy-forms==1.13.0
django-dotenv==1.4.2
django-extensions==3.2.0
django-input-mask==2.0.2
flake8==4.0.1
gunicorn==20.1.0
importlib-metadata==4.2.0
mccabe==0.6.1
Pillow==8.4.0
psycopg2==2.7.7
pycodestyle==2.8.0
pyflakes==2.4.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.1.1
whitenoise==5.3.0
zipp==3.6.0
